+++ SignNow.Net.Examples/Documents/GetTheDocumentHistory.cs
@@ -1,23 +1,33 @@
-using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
-using SignNow.Net.Model;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace SignNow.Net.Examples.Documents
+namespace SignNow.Net.Examples
 {
-    public static partial class DocumentExamples
+    public partial class DocumentExamples
     {
-        /// <summary>
-        /// Get Document History example
-        /// </summary>
-        /// <param name="documentId">Identity of the document</param>
-        /// <param name="signNowContext">signNow container with services.</param>
-        /// <returns><see cref="DocumentHistoryResponse"/></returns>
-        public static async Task<IReadOnlyList<DocumentHistoryResponse>>
-            GetTheDocumentHistory(string documentId, SignNowContext signNowContext)
+        [TestMethod]
+        public async Task GetTheDocumentHistoryAsync()
         {
-            return await signNowContext.Documents
-                .GetDocumentHistoryAsync(documentId)
+            // upload a document with a signature field
+            await using var fileStream = File.OpenRead(PdfWithSignatureField);
+            var document = await testContext.Documents
+                .UploadDocumentWithFieldExtractAsync(fileStream, "GetTheDocumentHistory.pdf")
                 .ConfigureAwait(false);
+
+            // get the document history
+            var documentHistory = await testContext.Documents
+                .GetDocumentHistoryAsync(document?.Id)
+                .ConfigureAwait(false);
+
+            // check the document history
+            Assert.IsTrue(documentHistory.All(item => item.DocumentId == document?.Id));
+            Assert.IsTrue(documentHistory.Any(item => item.Origin == "original"));
+            Assert.IsTrue(documentHistory.All(item => item.Email == credentials.Login));
+
+            // Clean up
+            DeleteTestDocument(document?.Id);
         }
     }
 }