+++ SignNow.Net.Examples/Folders/GetAllFolders.cs
@@ -1,20 +1,28 @@
+using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SignNow.Net.Model;
 
-namespace SignNow.Net.Examples.Folders
+namespace SignNow.Net.Examples
 {
-    public static partial class FolderExamples
+    public partial class FolderExamples
     {
-        /// <summary>
-        /// Get all folders of a user.
-        /// </summary>
-        /// <param name="signNowContext">signNow container with services.</param>
-        /// <returns>Returns all information about user's folders.</returns>
-        public static async Task<SignNowFolders> GetAllFolders(SignNowContext signNowContext)
+        [TestMethod]
+        public async Task GetAllFoldersAsync()
         {
-            return await signNowContext.Folders
+            // get all folders
+            var folders = await testContext.Folders
                 .GetAllFoldersAsync()
                 .ConfigureAwait(false);
+
+            // check if the root folder contains the default folders
+            Assert.IsInstanceOfType(folders, typeof(SignNowFolders));
+            Assert.AreEqual("Root", folders.Name);
+            Assert.IsTrue(folders.SystemFolder);
+
+            Assert.IsTrue(folders.Folders.Any(f => f.Name == "Documents"));
+            Assert.IsTrue(folders.Folders.Any(f => f.Name == "Archive"));
+            Assert.IsTrue(folders.Folders.Any(f => f.Name == "Templates"));
         }
     }
 }