+++ SignNow.Net.Examples/Users/ChangeUserDetails.cs
@@ -1,43 +1,77 @@
+using System;
 using System.Threading.Tasks;
-using SignNow.Net.Model;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SignNow.Net.Model.Requests;
 
-namespace SignNow.Net.Examples.Users
+namespace SignNow.Net.Examples
 {
-    public static partial class UserExamples
+    public partial class UserExamples
     {
-        /// <summary>
-        /// Updates user information i.e. first name, last name
-        /// </summary>
-        /// <param name="firstname">new User firstname</param>
-        /// <param name="lastname">new User lastname</param>
-        /// <param name="oldPwd">Old User password</param>
-        /// <param name="pwd">New User password</param>
-        /// <param name="signNowContext">signNow container with services.</param>
-        /// <returns></returns>
-        public static async Task<UserUpdateResponse> ChangeUserDetails(string firstname, string lastname, string oldPwd, string pwd, SignNowContext signNowContext)
+        [TestMethod]
+        public async Task ChangeUserDetailsAsync()
         {
-            var userUpdateOptions = new UpdateUserOptions
+            // Get current user details
+            var currentUser = await testContext.Users
+                .GetCurrentUserAsync()
+                .ConfigureAwait(false);
+
+            Console.WriteLine("Current user is: '{0}' '{1}'", currentUser.FirstName, currentUser.LastName);
+
+            // Update user details
+            var update = new UpdateUserOptions
             {
-                FirstName = firstname,
-                LastName = lastname,
-                OldPassword = oldPwd,
-                Password = pwd
+                FirstName = "signNow",
+                LastName = currentUser.LastName + "Updated",
+                OldPassword = credentials.Password,
+                Password = credentials.Password,
+                LogOutAll = false
             };
 
-            return await signNowContext.Users.UpdateUserAsync(userUpdateOptions)
+            var updateUserDetails = await testContext.Users
+                .UpdateUserAsync(update)
                 .ConfigureAwait(false);
-        }
 
-        /// <summary>
-        /// Sends password reset link via email example
-        /// </summary>
-        /// <param name="email">User email</param>
-        /// <param name="signNowContext">signNow container with services.</param>
-        /// <returns></returns>
-        public static async Task SendsPasswordResetLink(string email, SignNowContext signNowContext)
-        {
-            await signNowContext.Users.SendPasswordResetLinkAsync(email)
+            // get updated user details
+            var updatedUser = await testContext.Users
+                .GetCurrentUserAsync()
+                .ConfigureAwait(false);
+
+            Console.WriteLine("Updated user is: '{0}' '{1}'", updatedUser.FirstName, updatedUser.LastName);
+
+            // check if user details were updated
+            Assert.AreEqual(currentUser.FirstName, updateUserDetails.FirstName);
+            Assert.AreNotEqual(currentUser.LastName, updateUserDetails.LastName);
+            Assert.AreEqual(currentUser.LastName + "Updated", updateUserDetails.LastName);
+
+            // Revert user details back
+            var revert = new UpdateUserOptions
+            {
+                FirstName = currentUser.FirstName,
+                LastName = currentUser.LastName,
+                OldPassword = credentials.Password,
+                Password = credentials.Password,
+                LogOutAll = false
+            };
+
+            var revertedUser = await testContext.Users
+                .UpdateUserAsync(revert)
+                .ConfigureAwait(false);
+
+            // get last user details
+            var lastUserInfo = await testContext.Users
+                .GetCurrentUserAsync()
+                .ConfigureAwait(false);
+
+            Console.WriteLine("Reverted user is: '{0}' '{1}'", lastUserInfo.FirstName, lastUserInfo.LastName);
+
+            // check if user details were reverted
+            Assert.AreEqual(currentUser.FirstName, lastUserInfo.FirstName);
+            Assert.AreEqual(currentUser.LastName, lastUserInfo.LastName);
+            Assert.AreEqual(currentUser.LastName, revertedUser.LastName);
+
+            // send user password reset email
+            await testContext.Users
+                .SendPasswordResetLinkAsync(currentUser.Email)
                 .ConfigureAwait(false);
         }
     }