+++ SignNow.Net/Model/Responses/CreateDocumentFromTemplateResponse.cs
@@ -1,16 +1,11 @@
-using Newtonsoft.Json;
+using SignNow.Net.Model.Responses.GenericResponses;
 
 namespace SignNow.Net.Model.Responses
 {
     /// <summary>
     /// Represents response from signNow API for Create Document from Template request.
     /// </summary>
-    public class CreateDocumentFromTemplateResponse
+    public class CreateDocumentFromTemplateResponse :IdResponse
     {
-        /// <summary>
-        /// Identity of new Document.
-        /// </summary>
-        [JsonProperty("id")]
-        public string Id { get; set; }
     }
 }