+++ SignNow.Net/Model/Responses/DocumentGroupCreateResponse.cs
@@ -1,0 +1,9 @@
+using SignNow.Net.Model.Responses.GenericResponses;
+
+namespace SignNow.Net.Model.Responses
+{
+    /// <summary>
+    /// Represents response from signNow API for Create Document Group request.
+    /// </summary>
+    public class DocumentGroupCreateResponse : IdResponse { }
+}